project/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==1.10.13
PyMuPDF==1.24.1
python-multipart==0.0.9
langchain-experimental==0.0.53
llamaapi==0.1.36
langchain==0.1.11
fastapi==0.109.2
crud==0.1
fitz==0.0.1.dev2
Django==4.2.7
djangorestframework==3.14.0
pillow==10.2.0